* type-jig.js: lookahead in chars, fix WPM calc.
WPM was counting the exercise length (including all of the
lookahead words that had been loaded). I fixed it to count the
characters that you actually typed, though it's still js chars, so
values that get encoded as two UTF-16 words will count as two
characters.

Also when checking the answer, I explicitly converted the exercise
"word" to a string. This will allow us to make exercises out of
steno objects that include the stroke(s).