Don't blank answer before counting chars (fix WPM)
A few lines after this line of code that blanked out the answer text, is a line
that checks how many characters it contains, which figures into the WPM
calculation.

This line (removed by this patch) was making that calculation come out wrong.
Also it appears to be redundant, since the entire contents of this element are
replaced in the last line in this function.