Added Live WPM Counter and a WPM graph!
I added a Live WPM Counter that is located right below the Timer.
There's a global option located below the Hint toggle that allows you to disable it.

Currently the graph uses Graph.js and uses the Live WPM object to log a WPM every second and display it. The graph averages out the first 5 seconds to minimize spikes.

The WPM currently hooks into the Timer funcitonality  and runs an on update function that i have hooked into the constructor of the Timer prototype.